add 'next' api
