+++ bin/next-dev
@@ -3,7 +3,6 @@ import { resolve, join } from 'path'
 import parseArgs from 'minimist'
 import { exists } from 'mz/fs'
 import Server from '../server'
-import clean from '../server/build/clean'
 
 const argv = parseArgs(process.argv.slice(2), {
   alias: {
@@ -18,10 +17,9 @@ const argv = parseArgs(process.argv.slice(2), {
 
 const dir = resolve(argv._[0] || '.')
 
-clean(dir)
+const srv = new Server(dir, { dev: true })
+srv.start(argv.port)
 .then(async () => {
-  const srv = new Server({ dir, dev: true, hotReload: true })
-  await srv.start(argv.port)
   console.log('> Ready on http://localhost:%d', argv.port)
 
   // Check if pages dir exists and warn if not
