+++ examples/custom-server/package.json
@@ -1,0 +1,8 @@
+{
+  "scripts": {
+    "start": "node server.js"
+  },
+  "dependencies": {
+    "next": "*"
+  }
+}
