+++ server/hot-reloader.js
@@ -2,12 +2,12 @@ import { join, relative, sep } from 'path'
 import webpackDevMiddleware from 'webpack-dev-middleware'
 import webpackHotMiddleware from 'webpack-hot-middleware'
 import webpack from './build/webpack'
+import clean from './build/clean'
 import read from './read'
 
 export default class HotReloader {
-  constructor (dir, dev = false) {
+  constructor (dir) {
     this.dir = dir
-    this.dev = dev
     this.middlewares = []
     this.webpackDevMiddleware = null
     this.webpackHotMiddleware = null
@@ -37,7 +37,10 @@ export default class HotReloader {
   }
 
   async prepareMiddlewares () {
-    const compiler = await webpack(this.dir, { hotReload: true, dev: this.dev })
+    const [compiler] = await Promise.all([
+      webpack(this.dir, { dev: true }),
+      clean(this.dir)
+    ])
 
     compiler.plugin('after-emit', (compilation, callback) => {
       const { assets } = compilation
