+++ server/next.js
@@ -1,0 +1,5 @@
+import App from './app'
+
+module.exports = (dir, opts) => {
+  return new App(dir, opts)
+}
